Getters: Return integers instead of doubles

+++ src/main/scala/org/widok/moment/Getters.scala
@@ -5,11 +5,11 @@ import scala.scalajs.js
 @js.native
 trait Getters extends js.Object {
   def millisecond(): Int = js.native
-  def second(): Double = js.native
-  def minute(): Double = js.native
-  def hour(): Double = js.native
-  def day(): Double = js.native
-  def month(): Double = js.native
-  def year(): Double = js.native
+  def second(): Int = js.native
+  def minute(): Int = js.native
+  def hour(): Int = js.native
+  def day(): Int = js.native
+  def month(): Int = js.native
+  def year(): Int = js.native
   def date(): Int = js.native
 }
